refactor: move URI method of BackendService to uri.go
