+++ internal/pkg/describe/backend_service.go
@@ -33,11 +33,11 @@ type BackendServiceDescriber struct {
 func NewBackendServiceDescriber(opt NewServiceConfig) (*BackendServiceDescriber, error) {
 	describer := &BackendServiceDescriber{
 		ecsServiceDescriber: &ecsServiceDescriber{
-			app:             opt.App,
-			svc:             opt.Svc,
-			enableResources: opt.EnableResources,
-			store:           opt.DeployStore,
-			svcStackDescriber:    make(map[string]ecsStackDescriber),
+			app:               opt.App,
+			svc:               opt.Svc,
+			enableResources:   opt.EnableResources,
+			store:             opt.DeployStore,
+			svcStackDescriber: make(map[string]ecsStackDescriber),
 		},
 
 		envDescriber: make(map[string]envDescriber),
@@ -68,32 +68,6 @@ func NewBackendServiceDescriber(opt NewServiceConfig) (*BackendServiceDescriber,
 		return nil
 	}
 	return describer, nil
-}
-
-// URI returns the service discovery namespace and is used to make
-// BackendServiceDescriber have the same signature as WebServiceDescriber.
-func (d *BackendServiceDescriber) URI(envName string) (string, error) {
-	if err := d.initDescribers(envName); err != nil {
-		return "", err
-	}
-	svcStackParams, err := d.svcStackDescriber[envName].Params()
-	if err != nil {
-		return "", fmt.Errorf("get stack parameters for environment %s: %w", envName, err)
-	}
-	port := svcStackParams[cfnstack.LBWebServiceContainerPortParamKey]
-	if port == cfnstack.NoExposedContainerPort {
-		return BlankServiceDiscoveryURI, nil
-	}
-	endpoint, err := d.envDescriber[envName].ServiceDiscoveryEndpoint()
-	if err != nil {
-		return "nil", fmt.Errorf("retrieve service discovery endpoint for environment %s: %w", envName, err)
-	}
-	s := serviceDiscovery{
-		Service:  d.svc,
-		Port:     port,
-		Endpoint: endpoint,
-	}
-	return s.String(), nil
 }
 
 // Describe returns info of a backend service.
